controller/src/old/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
PyYAML==6.0
requests==2.28.1
RPi.GPIO==0.7.1
RPLCD==1.3.0
smbus2==0.4.2
urllib3==1.26.13
